-- incremental outlook, -- fixed the bug when some frames erratically show on displays they don't belong

+++ Management/Controllers/MediaController.cs
@@ -251,7 +251,7 @@ namespace DisplayMonkey.Controllers
                 if (width <= 120 && height <= 120)
                 {
                     byte[] cache = HttpRuntime.Cache.GetOrAddAbsolute(
-                        string.Format("thumb_image_{0}_{1}x{2}_{3}", id, width, height, mode),
+                        string.Format("thumb_image_{0}_{1}x{2}_{3}", id, width, height, (int)mode),
                         () => {
                             byte [] img = db.Contents.Find(id).Data;
                             using (MemoryStream src = new MemoryStream(img))
